Ensure correct GUI updates after data change events
The current implementation directly adopts the subscription model
when the data_change callback is called from the ua stack.
However, this callback is executed from an ua thread.
Since qt gui applications are usually single threaded and perform all
their events in the main thread, a queued signal/slot connection is used
to ensure that all model updates are executed from the main thread.
As a result the new value reported by the data change event is displayed
instantly.